clean:save routine and alarm id

+++ app/src/main/java/com/rahim/data/alarm/AlarmSong.kt
@@ -21,6 +21,6 @@ interface AlarmSong {
 //
 //    // play ringtone
 //    ringtone.play()
-    fun playRingtone(context:Context,alarmId: Int?)
-    fun stopRingtone(ringtone: Ringtone?,context:Context,alarmId: Int?)
+    fun playRingtone(context:Context,alarmId: Long?)
+    fun stopRingtone(ringtone: Ringtone?,context:Context,alarmId: Long?)
 }
